getting closer to making UART connection work

+++ UART_communication/UART_communication.h
@@ -18,7 +18,7 @@ void init_clocks(void);
 void init_communication(void);
 
 //methods
-void send_data(uint16_t data, unsigned int nr_of_bytes);
+void send_data(uint8_t data, unsigned int nr_of_bytes);
 void send_data_pointer(unsigned char *data_pointer, unsigned char data_bytes_length);
